refactor(client): create app/ composition root
Move App.tsx, main.tsx, index.css, Layout.tsx, Sidebar.tsx and
NotificationCenter.tsx into src/app/. Lift the U2 auth-port install and
U3 logout-teardown subscription out of main.tsx into app/session.ts,
imported for its side effect before createRoot. Update index.html's
entry script src accordingly. App.tsx's route table, lazy() and eager
imports are unchanged in structure/order (R12) -- only specifier
strings changed for the new depth.
