+++ client/index.html
@@ -16,6 +16,6 @@
   </head>
   <body class="bg-background text-foreground font-data antialiased">
     <div id="root"></div>
-    <script type="module" src="/src/main.tsx"></script>
+    <script type="module" src="/src/app/main.tsx"></script>
   </body>
 </html>
