+++ client/src/app/App.tsx
@@ -1,49 +1,49 @@
 import { lazy, Suspense, type ComponentType } from 'react';
 import { Routes, Route, Navigate } from 'react-router-dom';
-import Layout from './components/Layout';
-import ProtectedRoute from './components/ProtectedRoute';
-import ErrorBoundary from './shared/components/ErrorBoundary';
-import Login from './pages/Login';
-import PWAInstallPrompt from './shared/components/PWAInstallPrompt';
-import { useAuthStore } from './store/authStore';
+import Layout from './Layout';
+import ProtectedRoute from '../components/ProtectedRoute';
+import ErrorBoundary from '../shared/components/ErrorBoundary';
+import Login from '../pages/Login';
+import PWAInstallPrompt from '../shared/components/PWAInstallPrompt';
+import { useAuthStore } from '../store/authStore';
 
 // Eagerly loaded (critical path)
-import Dashboard from './pages/Dashboard';
-import POS from './pages/POS';
-import Inventory from './pages/Inventory';
+import Dashboard from '../pages/Dashboard';
+import POS from '../pages/POS';
+import Inventory from '../pages/Inventory';
 
 // Lazy loaded
-const BarcodeTools = lazy(() => import('./pages/BarcodeTools'));
-const Deliveries = lazy(() => import('./pages/Deliveries'));
-const SalesHistory = lazy(() => import('./pages/SalesHistory'));
-const UsersPage = lazy(() => import('./pages/Users'));
-const CustomersPage = lazy(() => import('./pages/Customers'));
-const DistributorsPage = lazy(() => import('./pages/Distributors'));
-const CategoriesPage = lazy(() => import('./pages/Categories'));
-const SettingsPage = lazy(() => import('./pages/Settings'));
-const PurchaseOrdersPage = lazy(() => import('./pages/PurchaseOrders'));
-const AuditLogPage = lazy(() => import('./pages/AuditLog'));
-const PromotionsPage = lazy(() => import('./pages/Promotions'));
-const GiftCardsPage = lazy(() => import('./pages/GiftCards'));
-const StockCountPage = lazy(() => import('./pages/StockCount'));
-const ExportsPage = lazy(() => import('./pages/Exports'));
-const CustomerDisplay = lazy(() => import('./pages/CustomerDisplay'));
-const RegisterPage = lazy(() => import('./pages/Register'));
-const ShiftsPage = lazy(() => import('./pages/Shifts'));
-const ExpensesPage = lazy(() => import('./pages/Expenses'));
-const SegmentsPage = lazy(() => import('./pages/Segments'));
-const LayawayPage = lazy(() => import('./pages/Layaway'));
-const BundlesPage = lazy(() => import('./pages/Bundles'));
-const FeedbackPage = lazy(() => import('./pages/Feedback'));
-const BackupPage = lazy(() => import('./pages/Backup'));
-const BranchesPage = lazy(() => import('./pages/Branches'));
-const StorefrontPage = lazy(() => import('./pages/Storefront'));
-const OnlineOrdersPage = lazy(() => import('./pages/OnlineOrders'));
-const ReportBuilderPage = lazy(() => import('./pages/ReportBuilder'));
-const VendorsPage = lazy(() => import('./pages/Vendors'));
-const SmartPricingPage = lazy(() => import('./pages/SmartPricing'));
-const AiInsightsPage = lazy(() => import('./pages/AiInsights'));
-const AdvancedAnalyticsPage = lazy(() => import('./pages/AdvancedAnalytics'));
+const BarcodeTools = lazy(() => import('../pages/BarcodeTools'));
+const Deliveries = lazy(() => import('../pages/Deliveries'));
+const SalesHistory = lazy(() => import('../pages/SalesHistory'));
+const UsersPage = lazy(() => import('../pages/Users'));
+const CustomersPage = lazy(() => import('../pages/Customers'));
+const DistributorsPage = lazy(() => import('../pages/Distributors'));
+const CategoriesPage = lazy(() => import('../pages/Categories'));
+const SettingsPage = lazy(() => import('../pages/Settings'));
+const PurchaseOrdersPage = lazy(() => import('../pages/PurchaseOrders'));
+const AuditLogPage = lazy(() => import('../pages/AuditLog'));
+const PromotionsPage = lazy(() => import('../pages/Promotions'));
+const GiftCardsPage = lazy(() => import('../pages/GiftCards'));
+const StockCountPage = lazy(() => import('../pages/StockCount'));
+const ExportsPage = lazy(() => import('../pages/Exports'));
+const CustomerDisplay = lazy(() => import('../pages/CustomerDisplay'));
+const RegisterPage = lazy(() => import('../pages/Register'));
+const ShiftsPage = lazy(() => import('../pages/Shifts'));
+const ExpensesPage = lazy(() => import('../pages/Expenses'));
+const SegmentsPage = lazy(() => import('../pages/Segments'));
+const LayawayPage = lazy(() => import('../pages/Layaway'));
+const BundlesPage = lazy(() => import('../pages/Bundles'));
+const FeedbackPage = lazy(() => import('../pages/Feedback'));
+const BackupPage = lazy(() => import('../pages/Backup'));
+const BranchesPage = lazy(() => import('../pages/Branches'));
+const StorefrontPage = lazy(() => import('../pages/Storefront'));
+const OnlineOrdersPage = lazy(() => import('../pages/OnlineOrders'));
+const ReportBuilderPage = lazy(() => import('../pages/ReportBuilder'));
+const VendorsPage = lazy(() => import('../pages/Vendors'));
+const SmartPricingPage = lazy(() => import('../pages/SmartPricing'));
+const AiInsightsPage = lazy(() => import('../pages/AiInsights'));
+const AdvancedAnalyticsPage = lazy(() => import('../pages/AdvancedAnalytics'));
 
 type UserRole = 'Admin' | 'Cashier' | 'Delivery';
